requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
asn1crypto==1.5.1
backports.zoneinfo;python_version<"3.10"
beautifulsoup4==4.10.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cssselect2==0.6.0
cssutils==2.4.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1
django-admin-sortable2==1.0.4
django-auditlog==1.0.0
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-colorfield==0.6.3
django-currencies==0.10.1
django-debug-toolbar==3.2.4
django-easy-select2==1.5.8
django-import-export==2.7.1
django-inlinecss==0.3.0
django-js-asset==2.0.0
django-jsonfield-backport==1.0.5
django-modeltranslation==0.17.5
django-mptt==0.13.4
django-rosetta==0.9.8
django-weasyprint==2.1.0
django-wkhtmltopdf==3.4.0
et-xmlfile==1.1.0
fonttools==4.33.3
fpdf==1.7.2
future==0.18.2
greenlet==1.1.2
html2image==2.0.1
html5lib==1.1
idna==3.3
jsonfield==2.1.1
lxml==4.8.0
MarkupPy==1.14
msgpack==1.0.3
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
Pillow==9.0.1
polib==1.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pynliner==0.8.0
pynvim==0.4.3
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-pdf==0.39
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
singlemodeladmin==0.9
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.3
uritools==4.0.0
urllib3==1.26.8
weasyprint==55.0
webencodings==0.5.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
zopfli==0.2.1
